[cmake] Add variables so that msbuild find NumPy
On Windows, the GNU Scientific Library gets installed in a specific
folder. So for msbuild.exe to find it, we add more variables that can be
used when invoking cmake. For example:

cmake -D NUMPY_INCLUDE_DIR="C:\Anaconda2\Lib\site-packages\numpy\core\include" ...

This is needed because of error message `numpyconfig.h: no such file or
directory`